filter code to getBalanceChangeSummary
