+++ apps/explorer/src/pages/transaction-result/transaction-summary/BalanceChanges.tsx
@@ -7,9 +7,9 @@ import {
 	useCoinMetadata,
 	type BalanceChange,
 	useResolveSuiNSName,
+	getRecognizedUnRecognizedTokenChanges,
 } from '@mysten/core';
 import { Heading, Text } from '@mysten/ui';
-import { useMemo } from 'react';
 
 import { Banner } from '~/ui/Banner';
 import { Coin } from '~/ui/CoinsStack';
@@ -73,18 +73,8 @@ function BalanceChangeEntry({ change }: { change: BalanceChange }) {
 
 function BalanceChangeCard({ changes, owner }: { changes: BalanceChange[]; owner: string }) {
 	const { data: suinsDomainName } = useResolveSuiNSName(owner);
-	const { recognizedTokenChanges, unRecognizedTokenChanges } = useMemo(() => {
-		const recognizedTokenChanges = [];
-		const unRecognizedTokenChanges = [];
-		for (let change of changes) {
-			if (change.unRecognizedToken) {
-				unRecognizedTokenChanges.push(change);
-			} else {
-				recognizedTokenChanges.push(change);
-			}
-		}
-		return { recognizedTokenChanges, unRecognizedTokenChanges };
-	}, [changes]);
+	const { recognizedTokenChanges, unRecognizedTokenChanges } =
+		getRecognizedUnRecognizedTokenChanges(changes);
 
 	return (
 		<TransactionBlockCard
